added events on the hub, added prompt handlers from the clients, passed object data to the hub from client
