--- index.js
@@ -1,9 +1,0 @@
-'use strict';
-
-
-require('dotenv').config();
-
-const server = require('./lib/server.js');
-
-server.start(process.env.PORT || 3000);
-
