+++ package.json
@@ -2,7 +2,7 @@
   "name": "munchkinserver",
   "version": "1.0.0",
   "description": "backend server for munchkin game",
-  "main": "index.js",
+  "main": "hub.js",
   "scripts": {
     "test": "jest --coverage --verbose"
   },
@@ -23,6 +23,7 @@
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "find-config": "^1.0.0",
+    "inquirer": "^7.3.3",
     "socket.io": "^2.3.0",
     "socket.io-client": "^2.3.1"
   },
